ipcache: refactor IP cache change event into own type
Instead of passing a growing number of parameters to
IPIdentityMappingListener.OnIPIdentityCacheChange, move the parameters
into its own type ChangeEvent.

This allows to add/remove parameters to OnIPIdentityCacheChange without
having to change the IPIdentityMappingListener interface and all its
implementations. Also, this makes it easier to follow the meaning of
each parameter.

No functional changes.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>