cilium: retire neigh cache entries upon node events
Enforce retiring of neigh cache entries either on update or deletion of
new nodes. This will enforce the datapath to re-cache the neigh entry
from the fib. This is needed in order to not leave stale entries around.
Only do this when we have acceleration enabled right now since this is
equivalent to what the datapath does as well. We might later extend this
for tc as well.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>